# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: SaiLi AI API
  version: 0.2.0
servers:
  - url: http://localhost:8000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /api/v1/health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /api/v1/auth/register:
    post:
      summary: User registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Created
  /api/v1/auth/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Token
  /api/v1/auth/me:
    get:
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /api/v1/competitions:
    get:
      summary: List competitions
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: tag
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
  /api/v1/competitions/{competition_id}:
    get:
      summary: Competition detail
      parameters:
        - in: path
          name: competition_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /api/v1/favorites:
    get:
      summary: List favorites
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /api/v1/favorites/{competition_id}:
    post:
      summary: Add favorite
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: competition_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    delete:
      summary: Remove favorite
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: competition_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /api/v1/subscriptions:
    get:
      summary: List subscriptions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
    post:
      summary: Add subscription
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subscription_type, target]
              properties:
                subscription_type:
                  type: string
                  enum: [tag, competition]
                target:
                  type: string
      responses:
        '200':
          description: OK
    delete:
      summary: Remove subscription
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: subscription_type
          required: true
          schema:
            type: string
        - in: query
          name: target
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/reminders/settings:
    get:
      summary: List reminder settings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
    post:
      summary: Upsert reminder setting
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [days_before]
              properties:
                days_before:
                  type: integer
                enabled:
                  type: boolean
      responses:
        '200':
          description: OK
  /api/v1/recommendations:
    get:
      summary: List recommendations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /api/v1/admin/competitions:
    get:
      summary: Admin list competitions
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: OK
    post:
      summary: Admin create competition
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/admin/competitions/{competition_id}:
    put:
      summary: Admin update competition
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: competition_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/admin/competitions/{competition_id}/publish:
    post:
      summary: Publish competition
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: competition_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /api/v1/admin/competitions/{competition_id}/unpublish:
    post:
      summary: Unpublish competition
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: competition_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /api/v1/admin/ingest/source:
    post:
      summary: Trigger ingestion sync
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                source_id:
                  type: string
      responses:
        '200':
          description: OK
  /api/v1/admin/ingest/sources:
    get:
      summary: List ingestion source metrics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /api/v1/admin/config/api-providers:
    get:
      summary: Get API provider config
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
    put:
      summary: Update API provider config
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/admin/recommendation-rules:
    put:
      summary: Update recommendation rules
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                required: [key, weight]
                properties:
                  key:
                    type: string
                  weight:
                    type: integer
      responses:
        '200':
          description: OK
  /api/v1/admin/reminders/send:
    post:
      summary: Send due reminders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /api/v1/sample/hello:
    get:
      summary: Sample module greeting
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: hello from sample module